+++ backend/learn-persistence/src/main/java/com/prosper/learn/persistence/dataobject/RoadmapDO.java
@@ -6,23 +6,23 @@ import java.time.LocalDateTime;
 @Data
 public class RoadmapDO {
 
-    private int id;
+    private Long id;
 
     private String content;
 
     private String contentHash;
 
-    private int professionId;
+    private Long professionId;
 
-    private int creatorId;
+    private Long creatorId;
 
     private String description;
 
-    private int vote;
+    private Integer vote;
 
-    private int comment;
+    private Integer comment;
 
-    private double score;  // 计算出的排序分数
+    private Double score;  // 计算出的排序分数
 
     private LocalDateTime scoreCalculatedAt;  // 分数计算时间
